refactor: mongodb db분리 -> 단일db 구조로 수정
