+++ eatngo-infra/mongodb/src/main/kotlin/com/eatngo/mongo/config/MongoConfig.kt
@@ -1,11 +1,13 @@
 package com.eatngo.mongo.config
 
 import org.springframework.boot.autoconfigure.mongo.MongoProperties
+import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.context.annotation.Configuration
 
 /**
- * MongoProperties를 상속받아 Mongo 설정 값을 구성하는 설정 클래스
+ * Mongo DB 연결 정보를 구성하는 클래스
  */
 @Configuration
+@ConfigurationProperties(prefix = "datastore.mongodb")
 class MongoConfig: MongoProperties() {
 }
