--- eatngo-infra/mongodb/src/main/kotlin/com/eatngo/mongo/config/app/AppMongoConfig.kt
@@ -1,15 +1,0 @@
-package com.eatngo.mongo.config.app
-
-import com.eatngo.mongo.config.MongoConfig
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Primary
-
-/**
- * APP DB Config 클래스
- */
-@Primary
-@Configuration
-@ConfigurationProperties(prefix = "datastore.mongodb.app")
-class AppMongoConfig: MongoConfig() {
-}
